[update] refactoring util methods
