+++ trie.go
@@ -253,3 +253,14 @@ func getParameter(label string) string {
 
 	return label[leftI+1 : rightI]
 }
+
+// deleteEmpty removes an empty value in slice.
+func deleteEmpty(s []string) []string {
+	var r []string
+	for _, str := range s {
+		if str != "" {
+			r = append(r, str)
+		}
+	}
+	return r
+}
