--- utils.go
@@ -1,12 +1,0 @@
-package goblin
-
-// deleteEmpty removes an empty value in slice.
-func deleteEmpty(s []string) []string {
-	var r []string
-	for _, str := range s {
-		if str != "" {
-			r = append(r, str)
-		}
-	}
-	return r
-}
